[paging][mm] Fix issue #98
On systems where there are more than 4 GB of usable memory,
get_phys_mem_size() won't work as expected when Tilck is 32-bit,
simply because the function returns `ulong` instead of `u64` and
the 32-bit unsigned integer will wrap-around.

The solution requires chaning the return type of the function plus
a simple cast. However, this patch includes one extra change: it
clamps the value of get_phys_mem_size() to LINEAR_MAPPING_SIZE
(896 MB typically) in init_paging(). That's necessary to save heap
space used for ref-counting of physical pages that won't be used,
simply because Tilck doesn't support "high-mem".

+++ include/tilck/kernel/system_mmap.h
@@ -50,9 +50,9 @@ get_phys_mem_mb(void)
    return __mem_upper_kb >> 10;
 }
 
-static ALWAYS_INLINE ulong
+static ALWAYS_INLINE u64
 get_phys_mem_size(void)
 {
    extern u32 __mem_upper_kb;
-   return __mem_upper_kb << 10;
+   return ((u64)__mem_upper_kb) << 10;
 }
